backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-comment-migrate==0.1.7
django-cors-headers==3.13.0
django-filter==21.1
django-ranged-response==0.2.0
django-restql==0.15.3
django-simple-captcha==0.5.17
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
gevent==22.10.2
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.7
mysqlclient==2.1.1
numpy==1.26.1
openpyxl==3.0.10
packaging==23.2
pandas==2.1.2
Pillow==8.3.2
PyJWT==2.8.0
pyparsing==3.0.9
pyPEG2==2.15.2
pypinyin==0.48.0
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
redis==5.0.1
requests==2.28.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.16.0
whitenoise==6.6.0
# psycopg2==2.9.9
# uvicorn==0.23.2
# gunicorn==21.2.0
# gevent==23.9.1
# Pillow==10.1.0
dvadmin-celery==1.0.5
pyinstaller==6.8.0
